requirements.txt
Flask
RPi.GPIO
mfrc522
gunicorn
